requirements.txt
fastapi
uvicorn
langchain
python-dotenv
pydantic
requests
ollama
